add more tests and fix typo
