+++ pumpkin-world/src/generation/chunk_noise_router/density_function/test.rs
@@ -1,8 +1,15 @@
-use crate::noise_router::NOISE_ROUTER_ASTS;
+use pumpkin_util::assert_eq_delta;
+use serde::Deserialize;
+use std::fs;
+use std::path::Path;
+use std::sync::LazyLock;
 
-use super::{
-    ChunkNoiseFunction, ChunkNoiseFunctionWrapperHandler, NoisePos, ProtoChunkNoiseFunction,
-};
+use crate::noise_router::density_function_ast::DensityFunctionRepr;
+use crate::noise_router::NOISE_ROUTER_ASTS;
+use crate::read_data_from_file;
+
+use super::super::{ChunkNoiseFunction, ChunkNoiseFunctionWrapperHandler};
+use super::{NoisePos, ProtoChunkNoiseFunction};
 
 struct TestNoisePos {
     x: i32,
@@ -22,7 +29,7 @@ impl NoisePos for TestNoisePos {
     }
 }
 
-macro_rules! sample_function {
+macro_rules! sample_router_function {
     ($name:ident, $seed: expr, $pos: expr) => {{
         let function_ast = NOISE_ROUTER_ASTS.overworld().$name();
         let proto_function = ProtoChunkNoiseFunction::generate(function_ast, $seed);
@@ -48,45 +55,51 @@ fn test_normal_surface_noisified() {
 
     // TODO: Move these values to a file and create an extractor for them
     assert_eq!(
-        sample_function!(barrier_noise, seed, pos),
+        sample_router_function!(barrier_noise, seed, pos),
         -0.5400227274000677f64
     );
     assert_eq!(
-        sample_function!(fluid_level_floodedness_noise, seed, pos),
+        sample_router_function!(fluid_level_floodedness_noise, seed, pos),
         -0.4709571987777473f64
     );
     assert_eq!(
-        sample_function!(fluid_level_spread_noise, seed, pos),
+        sample_router_function!(fluid_level_spread_noise, seed, pos),
         -0.057269139961514365f64
     );
     assert_eq!(
-        sample_function!(lava_noise, seed, pos),
+        sample_router_function!(lava_noise, seed, pos),
         -0.16423603877333556f64
     );
     assert_eq!(
-        sample_function!(temperature, seed, pos),
+        sample_router_function!(temperature, seed, pos),
         0.1182379898645608f64
     );
     assert_eq!(
-        sample_function!(vegetation, seed, pos),
+        sample_router_function!(vegetation, seed, pos),
         -0.0013601677416915584f64
     );
     assert_eq!(
-        sample_function!(continents, seed, pos),
+        sample_router_function!(continents, seed, pos),
         -0.008171952121206487f64
     );
     assert_eq!(
-        sample_function!(erosion, seed, pos),
+        sample_router_function!(erosion, seed, pos),
         -0.10391073889243099f64
     );
-    assert_eq!(sample_function!(depth, seed, pos), 0.411882147192955f64);
-    assert_eq!(sample_function!(ridges, seed, pos), 0.011110323612534296f64);
     assert_eq!(
-        sample_function!(initial_density_without_jaggedness, seed, pos),
+        sample_router_function!(depth, seed, pos),
+        0.411882147192955f64
+    );
+    assert_eq!(
+        sample_router_function!(ridges, seed, pos),
+        0.011110323612534296f64
+    );
+    assert_eq!(
+        sample_router_function!(initial_density_without_jaggedness, seed, pos),
         7.668311608489972f64
     );
     assert_eq!(
-        sample_function!(final_density, seed, pos),
+        sample_router_function!(final_density, seed, pos),
         0.15719144891255343f64
     );
 
@@ -219,7 +232,7 @@ fn test_normal_surface_noisified() {
     ];
     for ((x, y, z), value) in values {
         let pos = TestNoisePos { x, y, z };
-        assert_eq!(sample_function!(vein_toggle, seed, pos), value);
+        assert_eq!(sample_router_function!(vein_toggle, seed, pos), value);
     }
 
     let values = [
@@ -351,7 +364,7 @@ fn test_normal_surface_noisified() {
     ];
     for ((x, y, z), value) in values {
         let pos = TestNoisePos { x, y, z };
-        assert_eq!(sample_function!(vein_ridged, seed, pos), value);
+        assert_eq!(sample_router_function!(vein_ridged, seed, pos), value);
     }
 
     let values = [
@@ -483,6 +496,163 @@ fn test_normal_surface_noisified() {
     ];
     for ((x, y, z), value) in values {
         let pos = TestNoisePos { x, y, z };
-        assert_eq!(sample_function!(vein_gap, seed, pos), value);
+        assert_eq!(sample_router_function!(vein_gap, seed, pos), value);
+    }
+}
+
+#[test]
+fn test_config_final_density() {
+    let expected_data: Vec<(i32, i32, i32, f64)> =
+        read_data_from_file!("../../../../assets/final_density_dump_7_4.json");
+
+    let function_ast = NOISE_ROUTER_ASTS.overworld().final_density();
+    let proto_function = ProtoChunkNoiseFunction::generate(function_ast, 0);
+    let mut function = ChunkNoiseFunction::new(
+        &proto_function,
+        ChunkNoiseFunctionWrapperHandler::TestNoiseConfig,
+    );
+
+    // This is a lot of data it iter over, one two skip a few done
+    for (x, y, z, sample) in expected_data.into_iter().step_by(5) {
+        let pos = TestNoisePos { x, y, z };
+        assert_eq_delta!(function.sample(&pos), sample, f64::EPSILON);
+    }
+}
+
+#[derive(Deserialize)]
+struct DensityFunctionReprs {
+    #[serde(rename = "overworld/base_3d_noise")]
+    base_3d_noise: DensityFunctionRepr,
+    #[serde(rename = "overworld/caves/spaghetti_2d_thickness_modulator")]
+    spaghetti_2d_thickness: DensityFunctionRepr,
+    #[serde(rename = "overworld/caves/pillars")]
+    cave_pillars: DensityFunctionRepr,
+    #[serde(rename = "overworld/caves/noodle")]
+    cave_noodle: DensityFunctionRepr,
+    #[serde(rename = "overworld/caves/spaghetti_roughness_function")]
+    spaghetti_roughness: DensityFunctionRepr,
+    #[serde(rename = "overworld/caves/entrances")]
+    cave_entrances: DensityFunctionRepr,
+}
+
+macro_rules! read_data_from_file_json5 {
+    ($path:expr) => {
+        serde_json5::from_str(
+            &fs::read_to_string(
+                Path::new(env!("CARGO_MANIFEST_DIR"))
+                    .parent()
+                    .unwrap()
+                    .join(file!())
+                    .parent()
+                    .unwrap()
+                    .join($path),
+            )
+            .expect("no data file"),
+        )
+        .expect("failed to decode data")
+    };
+}
+
+static DENSITY_FUNCTION_REPRS: LazyLock<DensityFunctionReprs> =
+    LazyLock::new(|| read_data_from_file_json5!("../../../../assets/density_function_tests.json"));
+
+#[test]
+fn test_base_3d_noise() {
+    let proto_function =
+        ProtoChunkNoiseFunction::generate(&DENSITY_FUNCTION_REPRS.base_3d_noise, 0);
+    let mut function = ChunkNoiseFunction::new(
+        &proto_function,
+        ChunkNoiseFunctionWrapperHandler::TestNoiseConfig,
+    );
+
+    let expected_data: Vec<(i32, i32, i32, f64)> =
+        read_data_from_file!("../../../../assets/converted_3d_overworld_7_4.json");
+    for (x, y, z, sample) in expected_data {
+        let pos = TestNoisePos { x, y, z };
+        assert_eq_delta!(function.sample(&pos), sample, f64::EPSILON);
+    }
+}
+
+#[test]
+fn test_base_cave_entrances() {
+    let proto_function =
+        ProtoChunkNoiseFunction::generate(&DENSITY_FUNCTION_REPRS.cave_entrances, 0);
+    let mut function = ChunkNoiseFunction::new(
+        &proto_function,
+        ChunkNoiseFunctionWrapperHandler::TestNoiseConfig,
+    );
+
+    let expected_data: Vec<(i32, i32, i32, f64)> =
+        read_data_from_file!("../../../../assets/converted_cave_entrances_overworld_7_4.json");
+    for (x, y, z, sample) in expected_data {
+        let pos = TestNoisePos { x, y, z };
+        assert_eq_delta!(function.sample(&pos), sample, f64::EPSILON);
+    }
+}
+
+#[test]
+fn test_base_spahetti_roughness() {
+    let proto_function =
+        ProtoChunkNoiseFunction::generate(&DENSITY_FUNCTION_REPRS.spaghetti_roughness, 0);
+    let mut function = ChunkNoiseFunction::new(
+        &proto_function,
+        ChunkNoiseFunctionWrapperHandler::TestNoiseConfig,
+    );
+
+    let expected_data: Vec<(i32, i32, i32, f64)> = read_data_from_file!(
+        "../../../../assets/converted_cave_spaghetti_rough_overworld_7_4.json"
+    );
+    for (x, y, z, sample) in expected_data {
+        let pos = TestNoisePos { x, y, z };
+        assert_eq_delta!(function.sample(&pos), sample, f64::EPSILON);
+    }
+}
+
+#[test]
+fn test_base_cave_noodle() {
+    let proto_function = ProtoChunkNoiseFunction::generate(&DENSITY_FUNCTION_REPRS.cave_noodle, 0);
+    let mut function = ChunkNoiseFunction::new(
+        &proto_function,
+        ChunkNoiseFunctionWrapperHandler::TestNoiseConfig,
+    );
+
+    let expected_data: Vec<(i32, i32, i32, f64)> =
+        read_data_from_file!("../../../../assets/converted_cave_noodle_7_4.json");
+    for (x, y, z, sample) in expected_data {
+        let pos = TestNoisePos { x, y, z };
+        assert_eq_delta!(function.sample(&pos), sample, f64::EPSILON);
+    }
+}
+
+#[test]
+fn test_base_cave_pillars() {
+    let proto_function = ProtoChunkNoiseFunction::generate(&DENSITY_FUNCTION_REPRS.cave_pillars, 0);
+    let mut function = ChunkNoiseFunction::new(
+        &proto_function,
+        ChunkNoiseFunctionWrapperHandler::TestNoiseConfig,
+    );
+
+    let expected_data: Vec<(i32, i32, i32, f64)> =
+        read_data_from_file!("../../../../assets/converted_cave_pillar_7_4.json");
+    for (x, y, z, sample) in expected_data {
+        let pos = TestNoisePos { x, y, z };
+        assert_eq_delta!(function.sample(&pos), sample, f64::EPSILON);
+    }
+}
+
+#[test]
+fn test_base_spaghetti_2d_thickness() {
+    let proto_function =
+        ProtoChunkNoiseFunction::generate(&DENSITY_FUNCTION_REPRS.spaghetti_2d_thickness, 0);
+    let mut function = ChunkNoiseFunction::new(
+        &proto_function,
+        ChunkNoiseFunctionWrapperHandler::TestNoiseConfig,
+    );
+
+    let expected_data: Vec<(i32, i32, i32, f64)> =
+        read_data_from_file!("../../../../assets/converted_cave_spaghetti_2d_thicc_7_4.json");
+    for (x, y, z, sample) in expected_data {
+        let pos = TestNoisePos { x, y, z };
+        assert_eq_delta!(function.sample(&pos), sample, f64::EPSILON);
     }
 }
